Policy Evento, Cada editor tendra su propios eventos

+++ app/Http/Controllers/EventoController.php
@@ -8,22 +8,11 @@ use Illuminate\Support\Facades\Storage;
 
 class EventoController extends Controller
 {
-
-  public function owner()
-    {
-      
-    }
-
-  public function view()
-  {
-      
-  }
-
-
     public function index()
     {
+      //$eventos = Evento::where('editor_id', auth()->guard('editor')->user()->id)->get()->all();
       $eventos = Evento::all();
-      //$this->authorize('pass', $eventos);
+      //$this->authorize('owner', $eventos);
       
       $editor_id = \Auth::guard('editor')->user()->id;
       $admin = Editor::where('id', $editor_id)->first();
@@ -47,9 +36,11 @@ class EventoController extends Controller
             'asistentes' => 'required',
             'imagen' => 'required',
           ]);
-
-        $entrada=$request->all();
-
+          $request->editor_id = \Auth::guard('editor')->user()->id;
+          //dd($request->editor_id);
+          
+          $entrada=$request->all();
+          $entrada['editor_id']= \Auth::guard('editor')->user()->id;
         if($request->hasFile('imagen'))
         {
             $files = $request->file('imagen');
@@ -76,6 +67,7 @@ class EventoController extends Controller
 
     public function update(Request $request, $id)
     {
+     
         $entrada=$request->all();
         if($request->hasFile('imagen'))
         {
@@ -86,12 +78,17 @@ class EventoController extends Controller
         }
         
         $evento = Evento::findOrFail($id);
-
+        if (auth()->guard('editor')->user()->can('owner', $evento)){
         $evento->update($entrada);
+        
         return redirect()->route('evento.index')->with([
             'eventoActualizacion'=>'Has actualizado correctamente el evento ',
             'clase-alerta'=>'alert-info',
             ]);
+          }
+          else{
+            return view('layouts_editor.editorAuthorize');
+          }
     }
 
   /*  public function show($id)
@@ -103,7 +100,7 @@ class EventoController extends Controller
     public function edit($id)
     {
       $evento = Evento::findOrFail($id);
-     // $this->authorize('pass', $eventos);
+     // $this->authorize('owner', $evento);
       $editor_id = \Auth::guard('editor')->user()->id;
       $admin = Editor::where('id', $editor_id)->first();
       return view ('layouts_evento.eventoEdit', compact('evento', 'admin'));
@@ -112,19 +109,23 @@ class EventoController extends Controller
     public function destroy($id)
     {
         $evento = Evento::findOrFail($id);
-        
+        if (auth()->guard('editor')->user()->can('owner', $evento)){
         $evento->delete();
 
         return redirect()->route('editor.index')->with([
             'eventoDelete'=>'Has eliminado el evento correctamente ',
             'clase-alerta'=>'alert-danger',
             ]);
+        }
+        else
+        {
+          return view('layouts_editor.editorAuthorize');
+        }
     }
 
     public function deleteList()
     {
       $eventos = Evento::all();
-      //$this->authorize('pass', $eventos);
       
       $editor_id = \Auth::guard('editor')->user()->id;
       $admin = Editor::where('id', $editor_id)->first();
@@ -135,7 +136,9 @@ class EventoController extends Controller
 
     public function delete($id)
     {
+      
       $evento = Evento::findOrFail($id);
+     
       $editor_id = \Auth::guard('editor')->user()->id;
       $admin = Editor::where('id', $editor_id)->first();
       return view('layouts_evento.eventoShow',compact('evento', 'admin'));
